requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
bidict==0.21.2
colorama==0.4.4
dnspython==1.16.0
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
flask-navbar==1.0
Flask-SocketIO==5.1.0
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-python==4.5.2.54
packaging==20.9
pep517==0.10.0
Pillow==8.2.0
pip-tools==6.1.0
pluggy==0.13.1
py==1.10.0
pytest==6.2.4
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
toml==0.10.2
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
